added in UI Prompt and also fixed the two instances of loops
there was two instances of alert police and weapon fire dertection loops which I then consolidarted them into one loop.
I also went ahead and created a UI prompt that will function now as its supposed to for the police men to acept or decline the blip alert

If they accept it it will go ahead and clip a blip aswell.